storage: move some list tests to generic package
